% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Worlds Population}
\format{

}
\usage{
WorldPopulation
}
\description{
Worlds Population for every Country, Region, subregion, or area.
The collection of data from 1950 to 2020
}
\keyword{datasets}
